requirements.txt
fastapi==0.115.11           # Framework API
uvicorn==0.34.0             # ASGI server
chromadb==0.6.3             # Vector database cho RAG
psycopg2-binary==2.9.10     # Driver PostgreSQL
python-dotenv==1.0.1        # Đọc biến môi trường
langchain==0.3.21           # Hỗ trợ RAG
sentence-transformers==2.6.1  # Tạo embeddings (nếu dùng trong RAG)
openai==1.66.3              # Gọi API OpenAI (nếu dùng)
tiktoken==0.9.0             # Tokenizer cho OpenAI
rank-bm25==0.2.2            # Xếp hạng tài liệu cho RAG
pytesseract==0.3.13         # OCR với Tesseract
pdfplumber==0.11.5          # Trích xuất văn bản từ PDF
pdf2image==1.17.0           # Chuyển PDF thành hình ảnh cho OCR
PyMuPDF==1.25.3             # Đọc PDF (hỗ trợ cả văn bản và hình ảnh)
python-docx==1.1.2          # Đọc DOCX
pandas==2.2.3               # Đọc XLSX và xử lý dữ liệu
pillow==11.1.0              # Xử lý hình ảnh cho OCR
underthesea==6.8.4          # Xử lý tiếng Việt
numpy==2.2.3                # Hỗ trợ tính toán (cần bởi nhiều thư viện)
requests==2.32.3            # Gửi HTTP requests (nếu cần)
SQLAlchemy==2.0.39          # Hỗ trợ database (nếu dùng ngoài PostgreSQL)
httpx==0.28.1               # HTTP client bất đồng bộ